+++ src/i18n-l10n-editor/services/translations/ije-microsoft-translator.ts
@@ -14,8 +14,12 @@ export class IJEMicrosoftTranslator implements IJETranslation {
 
         let _languages = [];
 
+        if (text === undefined || text.length === 0) {
+            return { [language]: text };
+        }
+
         // we only need the language and not the file name
-        let split = language.split("_");
+        let split = language.split('_');
         if (split.length === 2 && split[0].length === 2) {
             language = split[0];
             // if (split[1].length === 2) {
@@ -29,10 +33,10 @@ export class IJEMicrosoftTranslator implements IJETranslation {
             //     language += `_${split[2].toLocaleLowerCase()}`;
             // }
         }
-        
+
         for (let l = 0; l < languages.length; ++l) {
             let lang = languages[l];
-            let split = lang.split("_");
+            let split = lang.split('_');
             if (split.length === 2 && split[0].length === 2) {
                 lang = split[0];
                 // if (split[1].length === 2) {
@@ -51,15 +55,33 @@ export class IJEMicrosoftTranslator implements IJETranslation {
             }
         }
 
+        // need to save the {text} so we can put them back afterwards
+        const _substitutes: string[] = [];
+        let place = 0;
+        while (true) {
+            if (text.indexOf("}", place) > text.indexOf("{", place)) {
+                _substitutes.push(text.substring(text.indexOf("{", place), text.indexOf("}", place) + 1));
+                place = text.indexOf("}", place) + 1;
+            } else {
+                break;
+            }
+        }
 
         const apiKey = IJEConfiguration.TRANSLATION_SERVICE_API_KEY;
         if (apiKey.length === 0) {
-            vscode.window.showErrorMessage("Your Microsoft API is blank. please update setting i18nJsonEditor.translationServiceApiKey");
+            vscode.window.showErrorMessage('Your Microsoft API Key is blank. please update setting i18nJsonEditor.translationServiceApiKey');
 
             return { [language]: text };
         }
 
-        _languages.forEach(async lang => {
+        const apiRegion = IJEConfiguration.TRANSLATION_SERVICE_API_REGION;
+        if (apiKey.length === 0) {
+            vscode.window.showErrorMessage('Your Microsoft API Region is blank. please update setting i18nJsonEditor.translationServiceApiRegion');
+
+            return { [language]: text };
+        }
+
+        await _languages.forEach(async lang => {
             try {
                 var response = await axios({
                     baseURL: endpoint + `/translate`,
@@ -67,7 +89,7 @@ export class IJEMicrosoftTranslator implements IJETranslation {
                     method: 'post',
                     headers: {
                         'Ocp-Apim-Subscription-Key': '4ee49f8bbe114a9aa255ceb16ba4c4fa',
-                        'Ocp-Apim-Subscription-Region': "australiaeast",
+                        'Ocp-Apim-Subscription-Region': apiRegion,
                         'Content-type': 'application/json'
                     },
                     params: {
@@ -77,30 +99,31 @@ export class IJEMicrosoftTranslator implements IJETranslation {
                     },
                     data: [
                         {
-                            Text: text
+                            text: text
                         }
                     ],
                     responseType: 'json'
                 });
-    
+
                 const data = response.data;
-    
+
                 if (data.length === 0) {
                     return { [language]: text };
                 }
-    
-                this.results = Object.assign(
+
+ 
+                const results = Object.assign(
                     {},
                     ...languages
                         .filter(l => l !== language)
                         .map(l => ({
                             [l]: data[0].translations.filter(t => {
                                 let to = t.to;
-                                if (to.indexOf("_") !== -1) {
-                                    to = t.to.substring(0,t.to.indexOf("_"));
+                                if (to.indexOf('_') !== -1) {
+                                    to = t.to.substring(0, t.to.indexOf('_'));
                                 }
-                                if (to.indexOf("-") !== -1) {
-                                    to = t.to.substring(0,t.to.indexOf("-"));
+                                if (to.indexOf('-') !== -1) {
+                                    to = t.to.substring(0, t.to.indexOf('-'));
                                 }
                                 if (l.indexOf(to) !== -1) {
                                     return t.text as string;
@@ -110,24 +133,36 @@ export class IJEMicrosoftTranslator implements IJETranslation {
                 );
 
                 languages
-                .filter(l => l !== language)
-                .forEach(l => {
-                    if (this.results[l]) {
-                        let r = this.results[l][0];
-                        if (r) {
-                            translation.languages[l] = r['text'];
+                    .filter(l => l !== language)
+                    .forEach(l => {
+                        if (results[l]) {
+                            let r = results[l][0];
+                            if (r) {
+                                let _text = r['text'];
+                                place = 0;
+                                _substitutes.flatMap(s => {
+                                    if (_text.indexOf("}", place) > _text.indexOf("{", place)) {
+                                        _text =  _text.substring(0, _text.indexOf("{", place) - 1) + s + _text.substring(_text.indexOf("}", place) + 1);
+                                        place = _text.indexOf("}", place) + 1;
+                                    }
+                
+                                });
+                                                translation.languages[l] = _text;
+                                this._manager.refreshDataTable();
+                                //const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
+                                //sleep(250);
+                            }
                         }
-                    }
-                });
-
-                this._manager.refreshDataTable();
+                    });
 
                 return this.results;
             } catch (e) {
                 let err = e as AxiosError;
                 let r = err.response as AxiosResponse;
-                vscode.window.showErrorMessage(`${e.toString()}\n${r.data}\nlangage ${lang}`);
-            }    
+                console.log(`${e.toString()}\n${r.data}\nlangage ${lang} - ${text}`);
+                vscode.window.showErrorMessage(`${e.toString()}\n${r.data}\nlangage ${lang} - ${text}`);
+            }
         });
     }
 }
+